Cumulative sum analysis on binary dataset.

The cusumr command takes in a binary vector of data and returns a dataframe of rolling cusum scores per event. The dataframe is indicative of compliance with the dataset. Users can choose to include a learning phase and reset options based on default values for upper and lower decision limits as well as acceptable rates of success and error.

For plotting see the cusumr_plot command.

cusumr(
  events_outcomes,
  acceptable_rate = 0.2,
  unacceptable_rate = 0.4,
  type1_error_rate = 0.1,
  type2_error_rate = 0.1,
  learning = TRUE,
  reset = TRUE
)

Arguments

events_outcomes
vector = [n_samples of 0s and 1s]. Binary Integer values. If intergers are not 0 or 1, then the FALSE alarms should be explicitly given.

acceptable_rate
numeric, default=0.2. Number between 0 and 1. Acceptable success rate of the process being monitored.

unacceptable_rate
numeric, default=0.24 Number between 0 and 1.

type1_error_rate
numeric, default=0.1. Number between 0 and 1. A false positive or type 1 error rate

type2_error_rate
numeric, default=0.2 Number between 0 and 1. A false negative or type 2 error rate

learning
boolean, optional (default=True). Whether to start from learning phase or from monitoring phase.

reset
boolean, optional (default=True). Whether to reset score when the score hits the decision limit in monitoring phase. If yes, the cusum score will start at zero again and restart monitoring.

References

The use of the Cusum technique in the assessment of trainee competence in new procedures. Int J Qual Health Care. 2000 Oct;12(5):433-8.

https://www.ncbi.nlm.nih.gov/pubmed/11079224

Cumulative sum (CUSUM) assessment and medical education: a square peg in a round hole. Anaesthesia, 2011, 66, pages 243-254.

https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1365-2044.2011.06692.x

An application of the learning curve cumulative summation test to evaluate training for endotracheal intubation in emergency medicine. Emerg Med J, 2015;32:291?..294.

https://www.ncbi.nlm.nih.gov/pubmed/24154942

Examples

df <- c(0,0,0,0,1,1,1,0,0,1,1,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1)
cusumr(df)

#>    number_events events_outcomes cusum_score learning_signal reset_signal
#> 1              1               0   0.0000000               1            0
#> 2              2               0   0.0000000               1            0
#> 3              3               0   0.0000000               1            0
#> 4              4               0   0.0000000               1            0
#> 5              5               1  -0.2933049               1            0
#> 6              6               1  -0.5866099               1            0
#> 7              7               1  -0.8799148               1            0
#> 8              8               0  -0.1732198               1            0
#> 9              9               0   0.0000000               1            0
#> 10            10               1  -0.2933049               1            0
#> 11            11               1  -0.5866099               1            0
#> 12            12               0   0.0000000               1            0
#> 13            13               0   0.0000000               1            0
#> 14            14               1  -0.2933049               1            0
#> 15            15               1  -0.5866099               1            0
#> 16            16               0   0.0000000               1            0
#> 17            17               0   0.0000000               1            0
#> 18            18               0   0.0000000               1            0
#> 19            19               0   0.0000000               1            0
#> 20            20               0   0.0000000               1            0
#> 21            21               0   0.0000000               1            0
#> 22            22               0   0.0000000               1            0
#> 23            23               0   0.0000000               1            0
#> 24            24               0   0.0000000               1            0
#> 25            25               0   0.0000000               1            0
#> 26            26               0   0.0000000               1            0
#> 27            27               0   0.0000000               1            0
#> 28            28               0   0.0000000               1            0
#> 29            29               1  -0.2933049               1            0
#> 30            30               1  -0.5866099               1            0
#> 31            31               1  -0.8799148               1            0
#> 32            32               1  -1.1732198               1            0
#> 33            33               1  -1.4665247               1            0
#> 34            34               1  -1.7598297               1            0
#> 35            35               1  -2.0531346               1            0
#> 36            36               1  -2.3464396               1            0
#> 37            37               1   0.0000000               0            0
#> 38            38               1   0.0000000               0            0
#> 39            39               1   0.0000000               0            0
#> 40            40               1   0.0000000               0            0
#> 41            41               1   0.0000000               0            0
#> 42            42               1   0.0000000               0            0
#> 43            43               1   0.0000000               0            0

cusumr(events_outcomes = df, learning = FALSE, reset = TRUE)

#>    number_events events_outcomes cusum_score learning_signal reset_signal
#> 1              1               0   0.7066951               0            0
#> 2              2               0   1.4133901               0            0
#> 3              3               0   2.1200852               0            0
#> 4              4               0   2.8267802               0            1
#> 5              5               1   0.0000000               0            0
#> 6              6               1   0.0000000               0            0
#> 7              7               1   0.0000000               0            0
#> 8              8               0   0.7066951               0            0
#> 9              9               0   1.4133901               0            0
#> 10            10               1   1.1200852               0            0
#> 11            11               1   0.8267802               0            0
#> 12            12               0   1.5334753               0            0
#> 13            13               0   2.2401703               0            1
#> 14            14               1   0.0000000               0            0
#> 15            15               1   0.0000000               0            0
#> 16            16               0   0.7066951               0            0
#> 17            17               0   1.4133901               0            0
#> 18            18               0   2.1200852               0            0
#> 19            19               0   2.8267802               0            1
#> 20            20               0   0.7066951               0            0
#> 21            21               0   1.4133901               0            0
#> 22            22               0   2.1200852               0            0
#> 23            23               0   2.8267802               0            1
#> 24            24               0   0.7066951               0            0
#> 25            25               0   1.4133901               0            0
#> 26            26               0   2.1200852               0            0
#> 27            27               0   2.8267802               0            1
#> 28            28               0   0.7066951               0            0
#> 29            29               1   0.4133901               0            0
#> 30            30               1   0.1200852               0            0
#> 31            31               1   0.0000000               0            0
#> 32            32               1   0.0000000               0            0
#> 33            33               1   0.0000000               0            0
#> 34            34               1   0.0000000               0            0
#> 35            35               1   0.0000000               0            0
#> 36            36               1   0.0000000               0            0
#> 37            37               1   0.0000000               0            0
#> 38            38               1   0.0000000               0            0
#> 39            39               1   0.0000000               0            0
#> 40            40               1   0.0000000               0            0
#> 41            41               1   0.0000000               0            0
#> 42            42               1   0.0000000               0            0
#> 43            43               1   0.0000000               0            0